Monitor: Fix incorrect marshalling for rF2Rules and fix FO detection to keep order as Rolling if it started that way.

+++ Monitor/rF2SMMonitor/rF2SMMonitor/rF2Data.cs
@@ -566,7 +566,7 @@ namespace rF2SMMonitor
     {
       // input only
       public rF2TrackRulesCommand mCommand;        // recommended action
-      public long mID;                             // slot ID if applicable
+      public int mID;                             // slot ID if applicable
       public double mET;                           // elapsed time that event occurred, if applicable
     }
